malditas vistas, en uno esta bien en otro sale muy largo

+++ app/src/main/java/com/example/deligov2/Beans/PlatilloSA.java
@@ -1,8 +1,9 @@
 package com.example.deligov2.Beans;
 
 public class PlatilloSA {
+
     private String nombre;
-    private double price;
+    private float price;
 
 
 }
